// src/app/api/model/album.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface Album { 
    id?: number;
    addedDate?: string;
    title?: string;
    releaseYear?: number;
    releaseDate?: string;
    publisher?: string;
    genre?: string;
    style?: string;
    releaseType?: string;
    reissue?: boolean;
    artist?: string;
    albumArtist?: string;
    fan?: boolean;
    country?: string;
    city?: string;
    rating?: number;
    favorite?: boolean;
    wishlist?: boolean;
    owned?: boolean;
    tino?: boolean;
    wire?: boolean;
    hidden?: boolean;
    videoUrl?: string;
    wikiUrl?: string;
    discogsUrl?: string;
    bandcampUrl?: string;
    rymUrl?: string;
}